Move merged environment interfaces and serialize funcs to platform
