+++ src/vs/workbench/contrib/terminal/common/environmentVariable.ts
@@ -5,62 +5,10 @@
 
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { Event } from 'vs/base/common/event';
-import { IProcessEnvironment } from 'vs/base/common/platform';
 import { ThemeIcon } from 'vs/platform/theme/common/themeService';
-import { VariableResolver } from 'vs/workbench/contrib/terminal/common/terminalEnvironment';
+import { IEnvironmentVariableCollection, IMergedEnvironmentVariableCollection } from 'vs/platform/terminal/common/environmentVariable';
 
 export const IEnvironmentVariableService = createDecorator<IEnvironmentVariableService>('environmentVariableService');
-
-export enum EnvironmentVariableMutatorType {
-	Replace = 1,
-	Append = 2,
-	Prepend = 3
-}
-
-export interface IEnvironmentVariableMutator {
-	readonly value: string;
-	readonly type: EnvironmentVariableMutatorType;
-}
-
-export interface IExtensionOwnedEnvironmentVariableMutator extends IEnvironmentVariableMutator {
-	readonly extensionIdentifier: string;
-}
-
-export interface IEnvironmentVariableCollection {
-	readonly map: ReadonlyMap<string, IEnvironmentVariableMutator>;
-}
-
-export interface IEnvironmentVariableCollectionWithPersistence extends IEnvironmentVariableCollection {
-	readonly persistent: boolean;
-}
-
-export interface IMergedEnvironmentVariableCollectionDiff {
-	added: ReadonlyMap<string, IExtensionOwnedEnvironmentVariableMutator[]>;
-	changed: ReadonlyMap<string, IExtensionOwnedEnvironmentVariableMutator[]>;
-	removed: ReadonlyMap<string, IExtensionOwnedEnvironmentVariableMutator[]>;
-}
-
-/**
- * Represents an environment variable collection that results from merging several collections
- * together.
- */
-export interface IMergedEnvironmentVariableCollection {
-	readonly collections: ReadonlyMap<string, IEnvironmentVariableCollection>;
-	readonly map: ReadonlyMap<string, IExtensionOwnedEnvironmentVariableMutator[]>;
-
-	/**
-	 * Applies this collection to a process environment.
-	 * @param variableResolver An optional function to use to resolve variables within the
-	 * environment values.
-	 */
-	applyToProcessEnvironment(env: IProcessEnvironment, variableResolver?: VariableResolver): Promise<void>;
-
-	/**
-	 * Generates a diff of this connection against another. Returns undefined if the collections are
-	 * the same.
-	 */
-	diff(other: IMergedEnvironmentVariableCollection): IMergedEnvironmentVariableCollectionDiff | undefined;
-}
 
 /**
  * Tracks and persists environment variable collections as defined by extensions.
@@ -97,11 +45,9 @@ export interface IEnvironmentVariableService {
 	delete(extensionIdentifier: string): void;
 }
 
-/** [variable, mutator] */
-export type ISerializableEnvironmentVariableCollection = [string, IEnvironmentVariableMutator][];
-
-/** [extension, collection] */
-export type ISerializableEnvironmentVariableCollections = [string, ISerializableEnvironmentVariableCollection][];
+export interface IEnvironmentVariableCollectionWithPersistence extends IEnvironmentVariableCollection {
+	readonly persistent: boolean;
+}
 
 export interface IEnvironmentVariableInfo {
 	readonly requiresAction: boolean;
